DC and tally integration

+++ src/main/java/com/steel/product/application/dto/TallyBillingInvoiceListDTO.java
@@ -34,7 +34,7 @@ public class TallyBillingInvoiceListDTO {
 	private String godown;
 	private String uom;
 	private BigDecimal gstPercentage;
-	private Float quantity;
+	private BigDecimal quantity;
 	//private BigDecimal basePrice;
 	//private BigDecimal additionalPrice;
 	//private BigDecimal packingRate;
